rename authservice to auth utility
